requirements.txt
django==4.2.11
python-dotenv==1.0.1
psycopg2==2.9.9
Pillow==10.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
djangorestframework==3.15.1